Remove * alleles from VCF
HaplotypeCaller reports spanning or overlapping deletions (* allele), which cause downstream modules to break.
See https://software.broadinstitute.org/gatk/documentation/article?id=11029
Because we use HaplotypeCaller on a given set of variants ("sufam" option), the * allele will be reported alongside a given variant as a multiallele locus (the * allele will be reported if there are reads with deletions spanning the variant region; such cases are rare). In this case, we can safely remove the * allele.