:bug: facets fix `--outPrefix` variable
In an explicit rule if the target name ends with a recognized suffix `$*` is set to the target name minus the suffix. Problem is in the updated script we use `.done`, which is not a recognized suffix, and `$*` is set to empty string. See [Automatic-Variables](https://www.gnu.org/software/make/manual/html_node/Automatic-Variables.html).